Test(green): allow_to_sum_numbers_with_decimals

+++ src/core/stringCalculator.ts
@@ -1,5 +1,5 @@
 const getNumbersIn = (theOperation: string) => (
-    theOperation.split(/[,\n]/).map(number => parseInt(number))
+    theOperation.split(/[,\n]/).map(number => parseFloat(number))
 )
 
 const sumAllNumbersIn = (theOperationToIterate: number[]) => {
